maint: Fix warnings on OSX 10.12.4 (clang 8)
* src/Makefile.am
* src/formatter/AdapterFormatterPlayer.cpp
* src/formatter/FormatterFocusManager.cpp
* src/formatter/FormatterFocusManager.h
* src/formatter/FormatterScheduler.cpp
* src/formatter/NclExecutionObject.cpp
* src/formatter/NclFormatterDeviceRegion.cpp
* src/formatter/NclPresentationEvent.cpp
* src/isdbt-tuner/MulticastProvider.h
* src/isdbt-tuner/Tuner.cpp
* src/main.cpp
* src/mb/Display.cpp
* src/mb/SDL2ffmpeg.cpp
* src/mb/SDLSurface.cpp
* src/mb/SDLSurface.h
* src/ncl/Descriptor.cpp
* src/ncl/FocusDecoration.cpp
* src/ncl/IntervalAnchor.cpp
* src/ncl/LayoutRegion.cpp
* src/player/AVPlayer.cpp
* src/player/NewVideoPlayer.cpp
* src/player/NewVideoPlayer.h
* src/player/Player.cpp
* src/util/functions.cpp

+++ src/formatter/AdapterFormatterPlayer.cpp
@@ -1184,7 +1184,7 @@ AdapterFormatterPlayer::prepareScope (double offset)
 
   if (offset > 0)
     {
-      player->setMediaTime (offset);
+      player->setMediaTime ( (guint32)offset);
     }
 }
